split up Quota module
