+++ lib/plausible/billing/feature.ex
@@ -132,7 +132,7 @@ defmodule Plausible.Billing.Feature do
       def check_availability(%Plausible.Auth.User{} = user) do
         cond do
           free?() -> :ok
-          __MODULE__ in Quota.allowed_features_for(user) -> :ok
+          __MODULE__ in Quota.Limits.allowed_features_for(user) -> :ok
           true -> {:error, :upgrade_required}
         end
       end
