+++ lib/plausible/site/memberships/accept_invitation.ex
@@ -26,7 +26,7 @@ defmodule Plausible.Site.Memberships.AcceptInvitation do
   @spec transfer_ownership(Site.t(), Auth.User.t()) ::
           {:ok, Site.Membership.t()}
           | {:error,
-             Billing.Quota.over_limits_error()
+             Billing.Quota.Limits.over_limits_error()
              | Ecto.Changeset.t()
              | :transfer_to_self
              | :no_plan}
@@ -54,7 +54,7 @@ defmodule Plausible.Site.Memberships.AcceptInvitation do
           {:ok, Site.Membership.t()}
           | {:error,
              :invitation_not_found
-             | Billing.Quota.over_limits_error()
+             | Billing.Quota.Limits.over_limits_error()
              | Ecto.Changeset.t()
              | :no_plan}
   def accept_invitation(invitation_id, user) do
