+++ lib/plausible_web/controllers/site/membership_controller.ex
@@ -29,8 +29,8 @@ defmodule PlausibleWeb.Site.MembershipController do
       |> Sites.get_for_user!(conn.assigns.site.domain)
       |> Plausible.Repo.preload(:owner)
 
-    limit = Plausible.Billing.Quota.team_member_limit(site.owner)
-    usage = Plausible.Billing.Quota.team_member_usage(site.owner)
+    limit = Plausible.Billing.Quota.Limits.team_member_limit(site.owner)
+    usage = Plausible.Billing.Quota.Usage.team_member_usage(site.owner)
     below_limit? = Plausible.Billing.Quota.below_limit?(usage, limit)
 
     render(
