+++ test/workers/check_usage_test.exs
@@ -31,8 +31,8 @@ defmodule Plausible.Workers.CheckUsageTest do
   test "does not send an email if account has been over the limit for one billing month", %{
     user: user
   } do
-    quota_stub =
-      Plausible.Billing.Quota
+    usage_stub =
+      Plausible.Billing.Quota.Usage
       |> stub(:monthly_pageview_usage, fn _user ->
         %{
           penultimate_cycle: %{date_range: @date_range, total: 9_000},
@@ -46,7 +46,7 @@ defmodule Plausible.Workers.CheckUsageTest do
       last_bill_date: Timex.shift(Timex.today(), days: -1)
     )
 
-    CheckUsage.perform(nil, quota_stub)
+    CheckUsage.perform(nil, usage_stub)
 
     assert_no_emails_delivered()
     assert Repo.reload(user).grace_period == nil
@@ -55,8 +55,8 @@ defmodule Plausible.Workers.CheckUsageTest do
   test "does not send an email if account is over the limit by less than 10%", %{
     user: user
   } do
-    quota_stub =
-      Plausible.Billing.Quota
+    usage_stub =
+      Plausible.Billing.Quota.Usage
       |> stub(:monthly_pageview_usage, fn _user ->
         %{
           penultimate_cycle: %{date_range: @date_range, total: 10_999},
@@ -70,7 +70,7 @@ defmodule Plausible.Workers.CheckUsageTest do
       last_bill_date: Timex.shift(Timex.today(), days: -1)
     )
 
-    CheckUsage.perform(nil, quota_stub)
+    CheckUsage.perform(nil, usage_stub)
 
     assert_no_emails_delivered()
     assert Repo.reload(user).grace_period == nil
@@ -79,8 +79,8 @@ defmodule Plausible.Workers.CheckUsageTest do
   test "sends an email when an account is over their limit for two consecutive billing months", %{
     user: user
   } do
-    quota_stub =
-      Plausible.Billing.Quota
+    usage_stub =
+      Plausible.Billing.Quota.Usage
       |> stub(:monthly_pageview_usage, fn _user ->
         %{
           penultimate_cycle: %{date_range: @date_range, total: 11_000},
@@ -94,7 +94,7 @@ defmodule Plausible.Workers.CheckUsageTest do
       last_bill_date: Timex.shift(Timex.today(), days: -1)
     )
 
-    CheckUsage.perform(nil, quota_stub)
+    CheckUsage.perform(nil, usage_stub)
 
     assert_email_delivered_with(
       to: [user],
@@ -107,8 +107,8 @@ defmodule Plausible.Workers.CheckUsageTest do
   test "sends an email suggesting enterprise plan when usage is greater than 10M ", %{
     user: user
   } do
-    quota_stub =
-      Plausible.Billing.Quota
+    usage_stub =
+      Plausible.Billing.Quota.Usage
       |> stub(:monthly_pageview_usage, fn _user ->
         %{
           penultimate_cycle: %{date_range: @date_range, total: 11_000_000},
@@ -122,7 +122,7 @@ defmodule Plausible.Workers.CheckUsageTest do
       last_bill_date: Timex.shift(Timex.today(), days: -1)
     )
 
-    CheckUsage.perform(nil, quota_stub)
+    CheckUsage.perform(nil, usage_stub)
 
     assert_delivered_email_matches(%{html_body: html_body})
 
@@ -136,8 +136,8 @@ defmodule Plausible.Workers.CheckUsageTest do
       |> Plausible.Auth.GracePeriod.start_changeset()
       |> Repo.update!()
 
-    quota_stub =
-      Plausible.Billing.Quota
+    usage_stub =
+      Plausible.Billing.Quota.Usage
       |> stub(:monthly_pageview_usage, fn _user ->
         %{
           penultimate_cycle: %{date_range: @date_range, total: 11_000},
@@ -151,7 +151,7 @@ defmodule Plausible.Workers.CheckUsageTest do
       last_bill_date: Timex.shift(Timex.today(), days: -1)
     )
 
-    CheckUsage.perform(nil, quota_stub)
+    CheckUsage.perform(nil, usage_stub)
 
     assert_no_emails_delivered()
     assert Repo.reload(user).grace_period.id == existing_grace_period.id
@@ -160,8 +160,8 @@ defmodule Plausible.Workers.CheckUsageTest do
   test "recommends a plan to upgrade to", %{
     user: user
   } do
-    quota_stub =
-      Plausible.Billing.Quota
+    usage_stub =
+      Plausible.Billing.Quota.Usage
       |> stub(:monthly_pageview_usage, fn _user ->
         %{
           penultimate_cycle: %{date_range: @date_range, total: 11_000},
@@ -175,7 +175,7 @@ defmodule Plausible.Workers.CheckUsageTest do
       last_bill_date: Timex.shift(Timex.today(), days: -1)
     )
 
-    CheckUsage.perform(nil, quota_stub)
+    CheckUsage.perform(nil, usage_stub)
 
     assert_delivered_email_matches(%{
       html_body: html_body
@@ -185,8 +185,8 @@ defmodule Plausible.Workers.CheckUsageTest do
   end
 
   test "clears grace period when plan is applicable again", %{user: user} do
-    quota_stub =
-      Plausible.Billing.Quota
+    usage_stub =
+      Plausible.Billing.Quota.Usage
       |> stub(:monthly_pageview_usage, fn _user ->
         %{
           penultimate_cycle: %{date_range: @date_range, total: 11_000},
@@ -200,11 +200,11 @@ defmodule Plausible.Workers.CheckUsageTest do
       last_bill_date: Timex.shift(Timex.today(), days: -1)
     )
 
-    CheckUsage.perform(nil, quota_stub)
+    CheckUsage.perform(nil, usage_stub)
     assert user |> Repo.reload() |> Plausible.Auth.GracePeriod.active?()
 
-    quota_stub =
-      Plausible.Billing.Quota
+    usage_stub =
+      Plausible.Billing.Quota.Usage
       |> stub(:monthly_pageview_usage, fn _user ->
         %{
           penultimate_cycle: %{date_range: @date_range, total: 11_000},
@@ -212,7 +212,7 @@ defmodule Plausible.Workers.CheckUsageTest do
         }
       end)
 
-    CheckUsage.perform(nil, quota_stub)
+    CheckUsage.perform(nil, usage_stub)
     refute user |> Repo.reload() |> Plausible.Auth.GracePeriod.active?()
   end
 
@@ -223,8 +223,8 @@ defmodule Plausible.Workers.CheckUsageTest do
         |> Plausible.Auth.GracePeriod.start_manual_lock_changeset()
         |> Repo.update!()
 
-      quota_stub =
-        Plausible.Billing.Quota
+      usage_stub =
+        Plausible.Billing.Quota.Usage
         |> stub(:monthly_pageview_usage, fn _user ->
           %{
             penultimate_cycle: %{date_range: @date_range, total: 1_100_000},
@@ -240,7 +240,7 @@ defmodule Plausible.Workers.CheckUsageTest do
         last_bill_date: Timex.shift(Timex.today(), days: -1)
       )
 
-      CheckUsage.perform(nil, quota_stub)
+      CheckUsage.perform(nil, usage_stub)
 
       assert_no_emails_delivered()
       assert Repo.reload(user).grace_period.id == existing_grace_period.id
@@ -250,8 +250,8 @@ defmodule Plausible.Workers.CheckUsageTest do
          %{
            user: user
          } do
-      quota_stub =
-        Plausible.Billing.Quota
+      usage_stub =
+        Plausible.Billing.Quota.Usage
         |> stub(:monthly_pageview_usage, fn _user ->
           %{
             penultimate_cycle: %{date_range: @date_range, total: 1_100_000},
@@ -267,7 +267,7 @@ defmodule Plausible.Workers.CheckUsageTest do
         last_bill_date: Timex.shift(Timex.today(), days: -1)
       )
 
-      CheckUsage.perform(nil, quota_stub)
+      CheckUsage.perform(nil, usage_stub)
 
       assert_email_delivered_with(
         to: [{nil, "enterprise@plausible.io"}],
@@ -279,8 +279,8 @@ defmodule Plausible.Workers.CheckUsageTest do
          %{
            user: user
          } do
-      quota_stub =
-        Plausible.Billing.Quota
+      usage_stub =
+        Plausible.Billing.Quota.Usage
         |> stub(:monthly_pageview_usage, fn _user ->
           %{
             penultimate_cycle: %{date_range: @date_range, total: 1},
@@ -300,7 +300,7 @@ defmodule Plausible.Workers.CheckUsageTest do
         last_bill_date: Timex.shift(Timex.today(), days: -1)
       )
 
-      CheckUsage.perform(nil, quota_stub)
+      CheckUsage.perform(nil, usage_stub)
 
       assert_email_delivered_with(
         to: [{nil, "enterprise@plausible.io"}],
@@ -309,8 +309,8 @@ defmodule Plausible.Workers.CheckUsageTest do
     end
 
     test "starts grace period when plan is outgrown", %{user: user} do
-      quota_stub =
-        Plausible.Billing.Quota
+      usage_stub =
+        Plausible.Billing.Quota.Usage
         |> stub(:monthly_pageview_usage, fn _user ->
           %{
             penultimate_cycle: %{date_range: @date_range, total: 1_100_000},
@@ -326,7 +326,7 @@ defmodule Plausible.Workers.CheckUsageTest do
         last_bill_date: Timex.shift(Timex.today(), days: -1)
       )
 
-      CheckUsage.perform(nil, quota_stub)
+      CheckUsage.perform(nil, usage_stub)
       assert user |> Repo.reload() |> Plausible.Auth.GracePeriod.active?()
     end
   end
@@ -335,8 +335,8 @@ defmodule Plausible.Workers.CheckUsageTest do
     test "checks usage one day after the last_bill_date", %{
       user: user
     } do
-      quota_stub =
-        Plausible.Billing.Quota
+      usage_stub =
+        Plausible.Billing.Quota.Usage
         |> stub(:monthly_pageview_usage, fn _user ->
           %{
             penultimate_cycle: %{date_range: @date_range, total: 11_000},
@@ -350,7 +350,7 @@ defmodule Plausible.Workers.CheckUsageTest do
         last_bill_date: Timex.shift(Timex.today(), days: -1)
       )
 
-      CheckUsage.perform(nil, quota_stub)
+      CheckUsage.perform(nil, usage_stub)
 
       assert_email_delivered_with(
         to: [user],
@@ -361,8 +361,8 @@ defmodule Plausible.Workers.CheckUsageTest do
     test "does not check exactly one month after last_bill_date", %{
       user: user
     } do
-      quota_stub =
-        Plausible.Billing.Quota
+      usage_stub =
+        Plausible.Billing.Quota.Usage
         |> stub(:monthly_pageview_usage, fn _user ->
           %{
             penultimate_cycle: %{date_range: @date_range, total: 11_000},
@@ -376,7 +376,7 @@ defmodule Plausible.Workers.CheckUsageTest do
         last_bill_date: ~D[2021-03-28]
       )
 
-      CheckUsage.perform(nil, quota_stub, ~D[2021-03-28])
+      CheckUsage.perform(nil, usage_stub, ~D[2021-03-28])
 
       assert_no_emails_delivered()
     end
@@ -385,8 +385,8 @@ defmodule Plausible.Workers.CheckUsageTest do
          %{
            user: user
          } do
-      quota_stub =
-        Plausible.Billing.Quota
+      usage_stub =
+        Plausible.Billing.Quota.Usage
         |> stub(:monthly_pageview_usage, fn _user ->
           %{
             penultimate_cycle: %{date_range: @date_range, total: 11_000},
@@ -400,7 +400,7 @@ defmodule Plausible.Workers.CheckUsageTest do
         last_bill_date: ~D[2021-06-29]
       )
 
-      CheckUsage.perform(nil, quota_stub, ~D[2021-08-30])
+      CheckUsage.perform(nil, usage_stub, ~D[2021-08-30])
 
       assert_email_delivered_with(
         to: [user],
